Fixed a broken unit test in the Mac build. Added 'report post' operation. Added missing serialization classes to some post operations. Bumped version.

+++ SnapperTests/Source/ModelTests.m
@@ -297,8 +297,8 @@
     STAssertNotNil(place, @"Unable to create place from JSON");
 
     STAssertTrue([place.factualId isEqualToString:@"19931850-dc2f-012e-561d-003048cad9da"], @"Place 'factualId' property should be equal to '19931850-dc2f-012e-561d-003048cad9da'");
-    STAssertTrue(place.longitude == -122.395012f, @"Place 'longitude' property should be equal to '-122.395012', is actually %f", place.longitude);
-    STAssertTrue(place.latitude == 37.776905f, @"Place 'latitude' property should be equal to '37.776905'; is actually %f", place.latitude);
+    STAssertTrue((float)place.longitude == -122.395012f, @"Place 'longitude' property should be equal to '-122.395012', is actually %f", place.longitude);
+    STAssertTrue((float)place.latitude == 37.776905f, @"Place 'latitude' property should be equal to '37.776905'; is actually %f", place.latitude);
     STAssertTrue(place.open == YES, @"Place 'open' property should be YES");
     STAssertTrue([place.categories count] == 1, @"Place 'categories' property should contain 1 item");
